U: Success! Success! Success!
We now have a the front-end with the images being displayed.

It was the file path for the images that was breaking it... I thought
that I would be like the `templates/navbar.html` whereby we have the app
name.

But in this case, no app name was needed as Django already knows that we
are inside it ( that's what she said )