U + B: Bug --> Moving Files and Folder Causes Trouble in Database
NOTE: This commit has no changes to the actual files and folders... We
just added the credentials for the new superuser that we are going to
create.

Now, moving the code for the `models.py` file from the 'landing' to the
'login' app crashes the whole database.

Therefore, I am going to be troubleshooting this tomorrow